Dockerfile: add file
Add the base Dockerfile to install a Zephyr development environment.
This includes:

* base package installation
* python requirements installation
* basic SDK installation without toolchain

For now, all apt packages recommended in the Zephyr documentation [1]
are included. A future upgrade may remove some non-required ones to make
the image lighter.

All Python requirements are installed as system packages to ease
administration. As the image should not include new packages in the
future, this is low-risk even if it bypasses the system administration
over the Python environment.

The image is usable (but not optimized for) distrobox. The 'west'
command can be exported to build Zephyr firmwares in this container from
the host.