zephyr-setup: add script and install in image
Add the zephyr-setup script, a shortcut to the SDK setup.sh to install
new toolchains. The script is installed in /usr/local/bin, and can be
exported to enable a distrobox user to install a new Zephyr toolchain
without needing to build/download a new container image with it
preinstalled.